Remove extra interface for TraceLocation and use the dataclass directly.
